+++ src/fractarchy.js
@@ -2022,8 +2022,24 @@ function Orbital (divContainer, data, quant, scale, ovalFillColor, ovalStrokeCol
         }, false);
     }
     
+    function setupWheelEvent () {
+        window.addEventListener('wheel', function (evt) {
+            event.deltaY * -0.01
+            magn = magn + event.deltaY * -0.0025;
+            if (magn < 1)
+                magn = 1;
+                
+            else if (magn > 1 / ratio)
+                magn = 1 / ratio;
+                    
+            rescale (magn);
+            redraw();
+        }, { passive: false });
+    }
+    
     setupMouseEvents ();
     setupTouchEvents ();
+    setupWheelEvent ();
 
     var initDone = false;
     divContainer.addEventListener('resize1', function (e) {